# Preliminary rework of the collection in the Schema abstraction