--- a/modules/core/src/smithy4s/http/internals/SchemaVisitorMetadataWriter.scala
+++ b/modules/core/src/smithy4s/http/internals/SchemaVisitorMetadataWriter.scala
@@ -22,6 +22,7 @@ import smithy4s.http.HttpBinding
 import smithy4s.http.internals.MetaEncode._
 import smithy4s.schema.Alt.SchemaAndValue
 import smithy4s.schema.{
+  CollectionTag,
   EnumValue,
   Field,
   Primitive,
@@ -84,26 +85,15 @@ object SchemaVisitorMetadataWriter extends SchemaVisitor[MetaEncode] { self =>
     }
   }
 
-  override def list[A](
+  override def collection[C[_], A](
       shapeId: ShapeId,
       hints: Hints,
+      tag: CollectionTag[C, A],
       member: Schema[A]
-  ): MetaEncode[List[A]] = {
+  ): MetaEncode[C[A]] = {
     self(member) match {
       case StringValueMetaEncode(f) =>
-        StringListMetaEncode[List[A]](listA => listA.map(f))
-      case _ => MetaEncode.empty
-    }
-  }
-
-  override def set[A](
-      shapeId: ShapeId,
-      hints: Hints,
-      member: Schema[A]
-  ): MetaEncode[Set[A]] = {
-    self(member) match {
-      case StringValueMetaEncode(f) =>
-        StringListMetaEncode[Set[A]](set => set.map(f).toList)
+        StringListMetaEncode[C[A]](c => tag.iterator(c).map(f).toList)
       case _ => MetaEncode.empty
     }
   }