--- a/modules/core/src/smithy4s/schema/SchemaVisitor.scala
+++ b/modules/core/src/smithy4s/schema/SchemaVisitor.scala
@@ -22,8 +22,7 @@ import Schema._
 // format: off
 trait SchemaVisitor[F[_]] extends (Schema ~> F) {
   def primitive[P](shapeId: ShapeId, hints: Hints, tag: Primitive[P]) : F[P]
-  def list[A](shapeId: ShapeId, hints: Hints, member: Schema[A]) : F[List[A]]
-  def set[A](shapeId: ShapeId, hints: Hints, member: Schema[A]): F[Set[A]]
+  def collection[C[_], A](shapeId: ShapeId, hints: Hints, tag: CollectionTag[C, A], member: Schema[A]): F[C[A]]
   def map[K, V](shapeId: ShapeId, hints: Hints, key: Schema[K], value: Schema[V]): F[Map[K, V]]
   def enumeration[E](shapeId: ShapeId, hints: Hints, values: List[EnumValue[E]], total: E => EnumValue[E]) : F[E]
   def struct[S](shapeId: ShapeId, hints: Hints, fields: Vector[SchemaField[S, _]], make: IndexedSeq[Any] => S) : F[S]
@@ -34,8 +33,7 @@ trait SchemaVisitor[F[_]] extends (Schema ~> F) {
 
   def apply[A](schema: Schema[A]) : F[A] = schema match {
     case PrimitiveSchema(shapeId, hints, tag) => primitive(shapeId, hints, tag)
-    case ListSchema(shapeId, hints, member) => list(shapeId, hints, member)
-    case SetSchema(shapeId, hints, member) => set(shapeId, hints, member)
+    case s: CollectionSchema[c, a] => collection[c,a](s.shapeId, s.hints, s.tag, s.member)
     case MapSchema(shapeId, hints, key, value) => map(shapeId, hints, key, value)
     case EnumerationSchema(shapeId, hints, values, total) => enumeration(shapeId, hints, values, total)
     case StructSchema(shapeId, hints, fields, make) => struct(shapeId, hints, fields, make)