--- a/modules/core/src/smithy4s/schema/SchematicRepr.scala
+++ b/modules/core/src/smithy4s/schema/SchematicRepr.scala
@@ -45,7 +45,7 @@ object SchematicRepr extends Schematic[Repr] {
 
   def document: String = "document"
 
-  def collection[C[_], S](tag: CollectionTag[C, S], fs: String): String =
+  def collection[C[_], S](tag: CollectionTag[C], fs: String): String =
     s"collection[$fs]"
 
   def uuid: String = "uuid"