… modules/core/test/src/smithy4s/ShapeIdHintsSmokeSpec.scala
… modules/core/test/src/smithy4s/http/internals/FromMetadataSchematic.scala
… modules/dynamic/test/src/smithy4s/dynamic/DefaultSchematic.scala
… modules/json/src/smithy4s/http/json/SchemaVisitorJCodec.scala