Replace list/set by a generic collection construct.