@Baccata 2022-07-05
@daddykotex please run the jsoniter benchmarks to see between the current version and this branch to whether this impacts the performance when decoding/encoding lists.