@plokhotnyuk 2022-07-05
> @daddykotex please run the jsoniter benchmarks to see between the current version and this branch to whether this impacts the performance when decoding/encoding lists.

It will impact performance - please, see `ListOfBooleanReading` and `ListOfBooleanWriting` benchmarks. 

Do we have in this new design an option to have specialized codecs for some collection tags?