@Baccata 2022-07-05
@plokhotnyuk if we keep the abstraction sealed, and expose the sealed members to recover the abstracted type, it'll allow for specialisation. I'd be fine with this.