@daddykotex 2022-07-05
Before:

```
[info] Benchmark                           (size)   Mode  Cnt         Score        Error  Units
[info] ListOfBooleansReading.smithy4sJson       1  thrpt    5  25993006.192 ± 720633.324  ops/s
[info] ListOfBooleansReading.smithy4sJson      10  thrpt    5   4430988.006 ±  34950.346  ops/s
[info] ListOfBooleansReading.smithy4sJson     100  thrpt    5    475538.740 ±   6568.466  ops/s
[info] ListOfBooleansReading.smithy4sJson    1000  thrpt    5     47875.703 ±    472.228  ops/s
[info] ListOfBooleansWriting.smithy4sJson       1  thrpt    5  67302416.332 ± 727037.113  ops/s
[info] ListOfBooleansWriting.smithy4sJson      10  thrpt    5  28217963.212 ± 828884.018  ops/s
[info] ListOfBooleansWriting.smithy4sJson     100  thrpt    5   3448066.045 ±  49790.724  ops/s
[info] ListOfBooleansWriting.smithy4sJson    1000  thrpt    5    283070.390 ±   2624.769  ops/s
```

After:

```
[info] Benchmark                           (size)   Mode  Cnt         Score         Error  Units
[info] ListOfBooleansReading.smithy4sJson       1  thrpt    5  23691087.074 ±  346083.797  ops/s
[info] ListOfBooleansReading.smithy4sJson      10  thrpt    5   3878777.112 ±  179290.385  ops/s
[info] ListOfBooleansReading.smithy4sJson     100  thrpt    5    428526.504 ±    3785.386  ops/s
[info] ListOfBooleansReading.smithy4sJson    1000  thrpt    5     42317.259 ±    1073.143  ops/s
[info] ListOfBooleansWriting.smithy4sJson       1  thrpt    5  70342211.467 ± 1301975.796  ops/s
[info] ListOfBooleansWriting.smithy4sJson      10  thrpt    5  28465901.292 ±  337870.882  ops/s
[info] ListOfBooleansWriting.smithy4sJson     100  thrpt    5   3409951.467 ±   49448.292  ops/s
[info] ListOfBooleansWriting.smithy4sJson    1000  thrpt    5    344767.207 ±    1870.146  ops/s
```

After specialization:
```
[info] Benchmark                           (size)   Mode  Cnt         Score         Error  Units
[info] ListOfBooleansReading.smithy4sJson       1  thrpt    5  25857027.983 ± 1667865.836  ops/s
[info] ListOfBooleansReading.smithy4sJson      10  thrpt    5   4439332.358 ±   34581.609  ops/s
[info] ListOfBooleansReading.smithy4sJson     100  thrpt    5    462419.385 ±    8232.583  ops/s
[info] ListOfBooleansReading.smithy4sJson    1000  thrpt    5     47899.194 ±     792.417  ops/s
[info] ListOfBooleansWriting.smithy4sJson       1  thrpt    5  67179957.102 ±  685996.973  ops/s
[info] ListOfBooleansWriting.smithy4sJson      10  thrpt    5  28097529.170 ±  923336.601  ops/s
[info] ListOfBooleansWriting.smithy4sJson     100  thrpt    5   3392081.900 ±   41404.484  ops/s
[info] ListOfBooleansWriting.smithy4sJson    1000  thrpt    5    282219.904 ±    6480.179  ops/s
```