@Baccata 2022-07-06
The changes seem to behave as expected. Using the Spark utility, we can estimate the size of instance on the heap.

```scala
//> using scala 2.13

import $ivy.`org.openjdk.jol:jol-core:0.16`
import $ivy.`org.apache.spark::spark-core:3.3.0`
import $ivy.`com.disneystreaming.smithy4s::smithy4s-json:dev`

import scala.reflect.ClassTag
import org.openjdk.jol.info.ClassLayout;
import scala.collection.immutable.ArraySeq
import org.apache.spark.util.SizeEstimator
import foo.Foo
import smithy4s.http.json._
import com.github.plokhotnyuk.jsoniter_scala.core._
import smithy4s.schema._

object Main {

  implicit val indexedSeqFooCodec = smithy4s.http.json.JCodec.fromSchema(Schema.indexedSeq(Foo.schema))
  implicit val indexedSeqIntCodec = smithy4s.http.json.JCodec.fromSchema(Schema.indexedSeq(Schema.int))

  def main(args: Array[String]): Unit =  {
    implicit val ct: ClassTag[Foo] = implicitly[ClassTag[Int]].asInstanceOf[ClassTag[Foo]]
    val naive = IndexedSeq.fill(1000)(1)
    val ints = ArraySeq.fill(1000)(1)
    val foos = ArraySeq.fill(1000)(Foo(1))
    val json = writeToArray[IndexedSeq[Int]](ints)
    val resultInt = readFromArray[IndexedSeq[Int]](json)
    val resultFoo = readFromArray[IndexedSeq[Foo]](json)
    println(SizeEstimator.estimate(naive)) // Ints are boxed in standard IndexedSeq construction, we get 4696
    println(SizeEstimator.estimate(ints)) // 4032
    println(SizeEstimator.estimate(foos)) // 4032
    println(SizeEstimator.estimate(resultInt)) // 4032
    println(SizeEstimator.estimate(resultFoo)) // 4032
  }
}

```