@daddykotex 2022-07-06
that's interesting, thanks for scala-cli script, TIL about the SizeEstimator