@Baccata 2022-07-07
I'm gonna go forward with this. We can refine later if need be.